Fix deletion of full rows
In the event of a collision of a tetromino whose bitmap had at least one
of the last 4 bits set (e.g. a vertically positioned type I tetromino),
when the last full row was the one containing the last block of that
tetromino and it was not the last row on the board, then that row was
not removed. This commit fixes that.